[C/PyTorch] Userbuffers and comm+GEMM overlap algorithms refactored and moved to TE/common (#1067)

See merge request dl/transformerengine/transformerengine!205

+++ transformer_engine/pytorch/csrc/extensions/cast.cu
@@ -95,7 +95,11 @@ std::vector<at::Tensor> fp8_cast_dbias(const at::Tensor& input, const at::Tensor
   void* scale_dptr = getDataPtr(scale, scale_offset);
   void* amax_dptr = getDataPtr(amax, amax_offset);
   void* scale_inv_dptr = getDataPtr(scale_inv, scale_inv_offset);
-  NVTEScalingMode nvte_scaling_mode = {scaling_mode[0], scaling_mode[1], scaling_mode[2]};
+  NVTEScalingMode nvte_scaling_mode = {
+      static_cast<int>(scaling_mode[0]),
+      static_cast<int>(scaling_mode[1]),
+      static_cast<int>(scaling_mode[2])
+  };
 
   auto input_cu = makeTransformerEngineTensor(input);
   auto dbias_cu = makeTransformerEngineTensor(grad_bias);
@@ -133,7 +137,11 @@ std::vector<at::Tensor> fp8_cast_dbias_dgelu(at::Tensor grad_output, at::Tensor 
   void* scale_dptr = getDataPtr(scale, scale_offset);
   void* amax_dptr = getDataPtr(amax, amax_offset);
   void* scale_inv_dptr = getDataPtr(scale_inv, scale_inv_offset);
-  NVTEScalingMode nvte_scaling_mode = {scaling_mode[0], scaling_mode[1], scaling_mode[2]};
+  NVTEScalingMode nvte_scaling_mode = {
+      static_cast<int>(scaling_mode[0]),
+      static_cast<int>(scaling_mode[1]),
+      static_cast<int>(scaling_mode[2])
+  };
 
   // Construct Transformer Engine tensors
   DType grad_output_type = GetTransformerEngineDType(grad_output.scalar_type());
@@ -178,7 +186,11 @@ std::vector<at::Tensor> fp8_cast_dbias_dsilu(at::Tensor grad_output, at::Tensor 
   void* scale_dptr = getDataPtr(scale, scale_offset);
   void* amax_dptr = getDataPtr(amax, amax_offset);
   void* scale_inv_dptr = getDataPtr(scale_inv, scale_inv_offset);
-  NVTEScalingMode nvte_scaling_mode = {scaling_mode[0], scaling_mode[1], scaling_mode[2]};
+  NVTEScalingMode nvte_scaling_mode = {
+      static_cast<int>(scaling_mode[0]),
+      static_cast<int>(scaling_mode[1]),
+      static_cast<int>(scaling_mode[2])
+  };
 
   // Construct Transformer Engine tensors
   DType grad_output_type = GetTransformerEngineDType(grad_output.scalar_type());
@@ -223,7 +235,11 @@ std::vector<at::Tensor> fp8_cast_dbias_drelu(at::Tensor grad_output, at::Tensor 
   void* scale_dptr = getDataPtr(scale, scale_offset);
   void* amax_dptr = getDataPtr(amax, amax_offset);
   void* scale_inv_dptr = getDataPtr(scale_inv, scale_inv_offset);
-  NVTEScalingMode nvte_scaling_mode = {scaling_mode[0], scaling_mode[1], scaling_mode[2]};
+  NVTEScalingMode nvte_scaling_mode = {
+      static_cast<int>(scaling_mode[0]),
+      static_cast<int>(scaling_mode[1]),
+      static_cast<int>(scaling_mode[2])
+  };
 
   // Construct Transformer Engine tensors
   DType grad_output_type = GetTransformerEngineDType(grad_output.scalar_type());
@@ -268,7 +284,11 @@ std::vector<at::Tensor> fp8_cast_dbias_dqgelu(at::Tensor grad_output, at::Tensor
   void* scale_dptr = getDataPtr(scale, scale_offset);
   void* amax_dptr = getDataPtr(amax, amax_offset);
   void* scale_inv_dptr = getDataPtr(scale_inv, scale_inv_offset);
-  NVTEScalingMode nvte_scaling_mode = {scaling_mode[0], scaling_mode[1], scaling_mode[2]};
+  NVTEScalingMode nvte_scaling_mode = {
+      static_cast<int>(scaling_mode[0]),
+      static_cast<int>(scaling_mode[1]),
+      static_cast<int>(scaling_mode[2])
+  };
 
   // Construct Transformer Engine tensors
   DType grad_output_type = GetTransformerEngineDType(grad_output.scalar_type());
@@ -313,7 +333,11 @@ std::vector<at::Tensor> fp8_cast_dbias_dsrelu(at::Tensor grad_output, at::Tensor
   void* scale_dptr = getDataPtr(scale, scale_offset);
   void* amax_dptr = getDataPtr(amax, amax_offset);
   void* scale_inv_dptr = getDataPtr(scale_inv, scale_inv_offset);
-  NVTEScalingMode nvte_scaling_mode = {scaling_mode[0], scaling_mode[1], scaling_mode[2]};
+  NVTEScalingMode nvte_scaling_mode = {
+      static_cast<int>(scaling_mode[0]),
+      static_cast<int>(scaling_mode[1]),
+      static_cast<int>(scaling_mode[2])
+  };
 
   // Construct Transformer Engine tensors
   DType grad_output_type = GetTransformerEngineDType(grad_output.scalar_type());
